+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 ### Modules
 
 - Escaped test run output before logging it, to avoid a rich ` MarkupError`
+- Add a new command `nf-core modules mulled` which can generate the name for a multi-tool container image.
 
 ## [v2.3.2 - Mercury Vulture Fixed Formatting](https://github.com/nf-core/tools/releases/tag/2.3.2) - [2022-03-24]